Remove QuickCSS; Implement Updater; asar bundle

+++ src/main/index.ts
@@ -3,10 +3,13 @@ import path from "path";
 import {execSync} from "child_process";
 import fs from "fs";
 
-const node_modules = path.resolve(__dirname, "..", "node_modules");
+const node_modules = path.resolve(
+    path.basename(__dirname) === "dist" ? path.resolve(__dirname, "..") : __dirname,
+    "node_modules"
+); 
 
 // Some people are incapable of running `pnpm install`.
-if (!fs.existsSync(node_modules)) {  
+if (!fs.existsSync(node_modules) && !__dirname.includes(".asar")) {  
     try {
         execSync("npm install --production", {
             cwd: __dirname
